+++ Sensors@claudiux/files/Sensors@claudiux/CHANGELOG.md
@@ -1,3 +1,9 @@
+### v3.8.1~20240901
+  * Fixes #6353. Repairs tooltip.
+
+### v3.8.0~20240818
+  * Fixes #6300. Better display on a vertical panel.
+
 ### v3.7.0~20240425
   * Added option to record values exceeding limits.
   * Fixes  #5606.