+++ sound150@claudiux/files/sound150@claudiux/CHANGELOG.md
@@ -1,3 +1,6 @@
+### v6.18.1~20240907
+  * Fixes #6301.
+
 ### v6.18.0~20240731
   * Improves certain icons.
   * Improved support for Pulse Effects and Easy Effects.